requirements.txt
absl-py==1.4.0
accelerate==1.0.1
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiohttp-retry==2.9.0
aioresponses==0.7.6
aiormq==6.4.2
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.9.1.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
awscli==1.35.12
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
boto3==1.35.46
botocore==1.35.46
Bottleneck==1.4.2
CacheControl==0.12.14
cachetools==5.5.0
catboost==1.2.7
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.7
cloudpickle==2.2.1
cmdstanpy==1.2.4
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
colorlog==6.8.2
comm==0.2.2
confluent-kafka==2.6.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dask==2022.10.2
databricks-cli==0.18.0
databricks-sdk==0.36.0
dataclasses-json==0.6.7
datasets==2.19.1
deap==1.4.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.3.0
docker==6.1.3
docopt==0.6.2
docutils==0.16
efficientnet==1.1.1
efficientnet_pytorch==0.7.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0
fastapi==0.104.1
fastjsonschema==2.20.0
fbmessenger==6.0.0
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
Flask==2.3.3
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
GPUtil==1.4.0
gradio==5.4.0
gradio_client==1.4.2
graphene==3.4.1
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
holidays==0.60
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.0
humanfriendly==10.0
idna==3.10
imageio==2.36.0
imbalanced-learn==0.12.4
importlib-metadata==6.11.0
importlib_resources==6.4.5
inference-cli==0.24.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonpickle==3.0.4
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
KTBoost==0.2.2
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.5.0
locket==1.0.0
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib==3.7.1
matplotlib-inline==0.1.7
mattermostwrapper==2.2
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mlflow-skinny==2.17.2
mpmath==1.3.0
msgpack==1.1.0
multidict==5.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
optree==0.13.0
optuna==4.0.0
optuna-integration==4.0.0
orjson==3.10.10
overrides==7.7.0
packaging==23.2
pamqp==3.2.1
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pendulum==3.0.0
Pillow==9.5.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pipdeptree==2.23.4
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
portalocker==2.10.1
prettytable==3.11.0
prometheus_client==0.21.0
prometheus_flask_exporter==0.23.1
prompt_toolkit==3.0.48
propcache==0.2.0
prophet==1.1.6
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.8
psycopg2-binary==2.9.10
PuLP==2.9.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
PyDrive==1.3.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.0.9
pyreadline3==3.5.4
pyrsistent==0.20.0
pytest==7.4.3
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-json-logger==2.0.7
python-multipart==0.0.12
python-socketio==5.11.4
python-version==0.0.2
pytorch-tabnet==4.1.0
pytz==2022.7.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
querystring-parser==1.2.4
questionary==1.10.0
randomname==0.1.5
redis==4.6.0
referencing==0.35.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rocketchat-API==1.30.0
rpds-py==0.20.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
ruff==0.7.1
s3transfer==0.10.3
safehttpx==0.1.1
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikeras==0.13.0
scikit-image==0.24.0
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sklearn-crfsuite==0.3.6
skypilot==0.5.0
slack_sdk==3.33.3
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
sqlparse==0.5.1
stack-data==0.6.3
stanio==0.5.1
starlette==0.27.0
statsmodels==0.14.4
stopit==1.1.2
structlog==23.3.0
structlog-sentry==2.1.0
supervision==0.22.0
sympy==1.13.1
tabulate==0.9.0
tarsafe==0.0.4
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-addons==0.18.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
time-machine==2.16.0
timm==1.0.11
tinycss2==1.3.0
tokenizers==0.20.1
tomli==2.0.2
tomlkit==0.12.0
toolz==1.0.0
torch==2.5.0
torchaudio==2.5.0
torchsummary==1.5.1
torchvision==0.20.0
tornado==6.4.1
TPOT==0.12.2
tqdm==4.66.5
traitlets==5.14.3
twilio==8.2.2
typeguard==2.13.3
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
ultralytics==8.3.23
ultralytics-thop==2.0.9
update-checker==0.18.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.11
uvicorn==0.24.0
waitress==2.1.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wordcloud==1.9.3
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.1.2
xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
